fix(backend): add DB rollback, fix monetary types to Decimal, improve health check
get_db() now rolls back on exception instead of leaving abandoned transactions.
Monetary fields use Decimal instead of float to prevent precision loss.
billing_service uses Decimal arithmetic with ROUND_HALF_UP.
Health endpoint now checks real DB connectivity.
Lifespan verifies DB connection on startup and disposes engine on shutdown.
CORS restricted to specific methods and headers.

+++ backend/app/models/billing.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from sqlalchemy import String, Integer, DateTime, ForeignKey, Numeric, Boolean, func, ARRAY
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
@@ -15,10 +16,10 @@ class BillingClaim(Base):
     procedure_code: Mapped[str] = mapped_column(String(10), nullable=False)
     modifier: Mapped[str] = mapped_column(String(5), default="U2")
     units: Mapped[int] = mapped_column(Integer, nullable=False)
-    gross_amount: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
-    platform_fee: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
-    pear_suite_fee: Mapped[float | None] = mapped_column(Numeric(10, 2))
-    net_payout: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    gross_amount: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
+    platform_fee: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
+    pear_suite_fee: Mapped[Decimal | None] = mapped_column(Numeric(10, 2))
+    net_payout: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     status: Mapped[str] = mapped_column(String(20), default="pending", index=True)
     pear_suite_claim_id: Mapped[str | None] = mapped_column(String(100))
     submitted_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
